build: 优化构建流程并添加新功能
- 在 HtmlBuiler 的 build 方法中添加了对 no_build_html 参数的判断，避免不必要的操作
- 更新了 main 函数中的参数解析器，增加了 --no-clean 参数
- 注释掉了 Builer 类的实例化，暂时禁用了其功能
- 更新了多个图标文件

+++ build.py
@@ -425,9 +425,11 @@ class HtmlBuiler:
 
     def build(self):
         os.makedirs("dist", exist_ok=True)
-        shutil.rmtree("html.old", ignore_errors=True)
-        if os.path.exists("html"):
-            shutil.copytree("html", "html.old")
+        
+        if not self.args.no_build_html:
+            shutil.rmtree("html.old", ignore_errors=True)
+            if os.path.exists("html"):
+                shutil.copytree("html", "html.old")
 
         print("Building HTML...")
         if not self.args.no_build_html:
@@ -458,8 +460,9 @@ if __name__ == "__main__":
     parser.add_argument("--no-zip", action="store_true", help="no zip")
     parser.add_argument("--no-check", action="store_true", help="no check")
     parser.add_argument("--default-version", help="default version", default="33.0.2")
+    parser.add_argument("--no-clean", action="store_true", help="no clean")
     args = parser.parse_args()
-    builders = [HtmlBuiler(args), Builer(args)]
+    builders = [HtmlBuiler(args),]# Builer(args)]
 
     for builder in builders:
         builder.build()
